* develop: (34 commits)
  Update tests to match new constructor pattern
  Reconfigure Canonical URL Resolver to better handle https requests
  Reconfigure UrlServiceProvider to support DI container
  Add Router instance to the IoC/DI container
  Adjust type hinting and visibility for path and page resolvers
  Add default request instance to IoC/DI container
  Change PackagePackCommand to a final class to prevent extension
  Fix: Make DatabaseSaver clear out child items when you save an item. Resolves #3129
  Add test for saving array over another array in both filesaver and databasesaver
  Adding upgrade
  Remove debugger
  Making constrain image work on backend; adding upgrade
  Fixes feature paragraph edit
  Fix symlinked core for *nix too
  Fix CLI run with symlinked core directory under Windows
  Allow manually setting radio id
  Apply urldecode before comparing two URLs
  Renaming the file variable to include in order to avoid variable name override
  add background-size and background-position options to Background design menu
  add date added to page drafts in Pages panel
  ...


Former-commit-id: 274a6e259c38812ca870d802937f5b7a18481e08
Former-commit-id: fe2424fe31f534638731d8adc573e33be98b31e9

+++ tests/tests/bootstrap.php
@@ -101,7 +101,13 @@ $config->set('concrete.cache.blocks', false);
 $config->set('concrete.cache.pages', false);
 $config->set('concrete.cache.enabled', false);
 
+/** @var Concrete\Core\Database\Connection\Connection $cn */
 $cn = $cms->make('database')->connection('travisWithoutDB');
+$cn->connect();
+if (!$cn->isConnected()) {
+    throw new \Exception('Unable to connect to test database, please create a user "travis" with no password with full privileges to a database "concrete5_tests"');
+}
+
 $cn->query('DROP DATABASE IF EXISTS concrete5_tests');
 $cn->query('CREATE DATABASE concrete5_tests');
 $cn->close();
